+++ MidCapERP.BusinessLogic/Repositories/CustomerAddressesBL.cs
@@ -53,13 +53,13 @@ namespace MidCapERP.BusinessLogic.Repositories
             return new JsonRepsonse<CustomerAddressesResponseDto>(dataTableFilterDto.Draw, customerAddressesData.TotalCount, customerAddressesData.TotalCount, customerAddressesData);
         }
 
-        public async Task<CustomerAddressesResponseDto> GetDetailsById(int Id, CancellationToken cancellationToken)
+        public async Task<CustomerAddressesResponseDto> GetDetailsById(Int64 Id, CancellationToken cancellationToken)
         {
             var data = await GetDetailsById(Id, cancellationToken);
             return _mapper.Map<CustomerAddressesResponseDto>(data);
         }
 
-        public async Task<CustomerAddressesRequestDto> GetById(int Id, CancellationToken cancellationToken)
+        public async Task<CustomerAddressesRequestDto> GetById(Int64 Id, CancellationToken cancellationToken)
         {
             var data = await CustomerAddressesGetById(Id, cancellationToken);
             return _mapper.Map<CustomerAddressesRequestDto>(data);
@@ -76,7 +76,7 @@ namespace MidCapERP.BusinessLogic.Repositories
             return _mapper.Map<CustomerAddressesRequestDto>(data);
         }
 
-        public async Task<CustomerAddressesRequestDto> UpdateCustomerAddresses(int Id, CustomerAddressesRequestDto model, CancellationToken cancellationToken)
+        public async Task<CustomerAddressesRequestDto> UpdateCustomerAddresses(Int64 Id, CustomerAddressesRequestDto model, CancellationToken cancellationToken)
         {
             var oldData = await CustomerAddressesGetById(Id, cancellationToken);
             UpdateCustomerAddresses(oldData);
@@ -85,7 +85,7 @@ namespace MidCapERP.BusinessLogic.Repositories
             return _mapper.Map<CustomerAddressesRequestDto>(data);
         }
 
-        public async Task<CustomerAddressesRequestDto> DeleteCustomerAddresses(int Id, CancellationToken cancellationToken)
+        public async Task<CustomerAddressesRequestDto> DeleteCustomerAddresses(Int64 Id, CancellationToken cancellationToken)
         {
             var customerAddressToInsert = await CustomerAddressesGetById(Id, cancellationToken);
             customerAddressToInsert.IsDeleted = true;
@@ -103,7 +103,7 @@ namespace MidCapERP.BusinessLogic.Repositories
             oldData.UpdatedUTCDate = DateTime.UtcNow;
         }
 
-        private async Task<CustomerAddresses> CustomerAddressesGetById(int Id, CancellationToken cancellationToken)
+        private async Task<CustomerAddresses> CustomerAddressesGetById(Int64 Id, CancellationToken cancellationToken)
         {
             var data = await _unitOfWorkDA.CustomerAddressesDA.GetById(Id, cancellationToken);
             if (data == null)
